+++ src/lib/Docs.svelte
@@ -6,12 +6,12 @@
 
 	import Breadcrumb from '$lib/Breadcrumb.svelte';
 	import {Tome, tomes_context} from '$lib/tome.js';
-	import Library_Primary_Nav from '$lib/Library_Primary_Nav.svelte';
-	import Library_Secondary_Nav from '$lib/Library_Secondary_Nav.svelte';
-	import Library_Tertiary_Nav from '$lib/Library_Tertiary_Nav.svelte';
+	import Docs_Primary_Nav from '$lib/Docs_Primary_Nav.svelte';
+	import Docs_Secondary_Nav from '$lib/Docs_Secondary_Nav.svelte';
+	import Docs_Tertiary_Nav from '$lib/Docs_Tertiary_Nav.svelte';
 	import Dialog from '$lib/Dialog.svelte';
-	import Library_Footer from '$lib/Library_Footer.svelte';
-	import {library_links_context} from '$lib/library_helpers.svelte.js';
+	import Docs_Footer from '$lib/Docs_Footer.svelte';
+	import {docs_links_context} from '$lib/docs_helpers.svelte.js';
 
 	interface Props {
 		tomes: Array<Tome>;
@@ -30,7 +30,7 @@
 	const TERTIARY_NAV_BREAKPOINT = 1000;
 	const SECONDARY_NAV_BREAKPOINT = 800;
 
-	const library_menu_width = '180px';
+	const docs_menu_width = '180px';
 
 	let show_secondary_nav_dialog = $state(false);
 	const toggle_secondary_nav_dialog = (show?: boolean): void => {
@@ -41,32 +41,32 @@
 		show_secondary_nav_dialog = false;
 	});
 
-	library_links_context.set();
+	docs_links_context.set();
 </script>
 
 <svelte:window onhashchange={() => (show_secondary_nav_dialog = false)} />
 
-<div class="library" style:--library_menu_width={library_menu_width}>
-	<Library_Primary_Nav {pkg} {breadcrumb_children}>
+<div class="library" style:--docs_menu_width={docs_menu_width}>
+	<Docs_Primary_Nav {pkg} {breadcrumb_children}>
 		<div class="nav_dialog_toggle">
 			<button class="plain" type="button" onclick={() => toggle_secondary_nav_dialog()}>menu</button
 			>
 		</div>
-	</Library_Primary_Nav>
+	</Docs_Primary_Nav>
 	<!-- TODO @many dialog navs -->
 	{#if !innerWidth.current || innerWidth.current > SECONDARY_NAV_BREAKPOINT}
 		<div class="secondary_nav_wrapper">
-			<Library_Secondary_Nav {tomes} />
+			<Docs_Secondary_Nav {tomes} />
 		</div>
 	{/if}
 	<main>
 		{@render children()}
 		<!-- TODO @many dialog navs -->
 		{#if !innerWidth.current || innerWidth.current > TERTIARY_NAV_BREAKPOINT}
-			<Library_Tertiary_Nav {tomes} {tomes_by_name} />
+			<Docs_Tertiary_Nav {tomes} {tomes_by_name} />
 		{/if}
 		<section class="box">
-			<Library_Footer {pkg}>
+			<Docs_Footer {pkg}>
 				<div class="mb_xl5">
 					<Breadcrumb>
 						{#if breadcrumb_children}
@@ -76,15 +76,15 @@
 						{/if}
 					</Breadcrumb>
 				</div>
-			</Library_Footer>
+			</Docs_Footer>
 		</section>
 	</main>
 </div>
 <!-- TODO @many dialog navs - instead of a dialog, probably use a popover (new component) -->
-<!-- TODO this is messy rendering `Library_Secondary_Nav` twice to handle responsive states with SSR correctly -->
+<!-- TODO this is messy rendering `Docs_Secondary_Nav` twice to handle responsive states with SSR correctly -->
 {#if show_secondary_nav_dialog && innerWidth.current && innerWidth.current <= TERTIARY_NAV_BREAKPOINT}
 	<Dialog onclose={() => (show_secondary_nav_dialog = false)}>
-		<div class="pane" style:--library_menu_width={library_menu_width}>
+		<div class="pane" style:--docs_menu_width={docs_menu_width}>
 			<div class="p_xl pb_0">
 				<Breadcrumb>
 					{#if breadcrumb_children}
@@ -95,8 +95,8 @@
 				</Breadcrumb>
 			</div>
 			<div class="px_lg pb_xl">
-				<Library_Secondary_Nav {tomes} sidebar={false} />
-				<Library_Tertiary_Nav {tomes} {tomes_by_name} sidebar={false} />
+				<Docs_Secondary_Nav {tomes} sidebar={false} />
+				<Docs_Tertiary_Nav {tomes} {tomes_by_name} sidebar={false} />
 			</div>
 		</div>
 	</Dialog>
@@ -104,24 +104,24 @@
 
 <style>
 	.library {
-		--library_primary_nav_height: 60px;
-		--library_secondary_nav_padding: var(--space_md); /* also used by the tertiary nav */
-		--library_content_padding: var(--space_xl5);
-		--library_content_max_width: calc(var(--width_md) + var(--library_content_padding) * 2);
+		--docs_primary_nav_height: 60px;
+		--docs_secondary_nav_padding: var(--space_md); /* also used by the tertiary nav */
+		--docs_content_padding: var(--space_xl5);
+		--docs_content_max_width: calc(var(--width_md) + var(--docs_content_padding) * 2);
 		/* the `+ 1px` solves some issue when scaling */
-		--library_sidebar_width: max(
-			calc(var(--library_menu_width) + 1px + var(--library_secondary_nav_padding) * 2),
-			calc((100% - var(--library_content_max_width)) / 2)
+		--docs_sidebar_width: max(
+			calc(var(--docs_menu_width) + 1px + var(--docs_secondary_nav_padding) * 2),
+			calc((100% - var(--docs_content_max_width)) / 2)
 		);
 		display: contents;
 	}
 
 	main {
 		position: relative;
-		min-height: calc(100vh - var(--library_primary_nav_height));
-		width: calc(100% - var(--library_sidebar_width) * 2);
-		max-width: var(--library_content_max_width);
-		padding: var(--library_content_padding);
+		min-height: calc(100vh - var(--docs_primary_nav_height));
+		width: calc(100% - var(--docs_sidebar_width) * 2);
+		max-width: var(--docs_content_max_width);
+		padding: var(--docs_content_padding);
 		padding-top: 0;
 		margin: 0 auto;
 		overflow: hidden; /* TODO maybe heavy handed */
@@ -140,12 +140,12 @@
 		border-radius: 0;
 	}
 
-	/* sync this breakpoint with `Library_Tertiary_Nav` and `Tome_Section_Header` */
+	/* sync this breakpoint with `Docs_Tertiary_Nav` and `Tome_Section_Header` */
 	@media (max-width: 1000px) {
 		main {
-			--library_content_padding: var(--space_xl);
-			/* handle the moved `Library_Tertiary_Nav` */
-			width: calc(100% - var(--library_sidebar_width));
+			--docs_content_padding: var(--space_xl);
+			/* handle the moved `Docs_Tertiary_Nav` */
+			width: calc(100% - var(--docs_sidebar_width));
 			margin-right: 0;
 		}
 
@@ -154,10 +154,10 @@
 		}
 	}
 
-	/* sync this breakpoint with `Library_Primary_Nav`, `Library_Secondary_Nav`, and `Tome_Section_Header` */
+	/* sync this breakpoint with `Docs_Primary_Nav`, `Docs_Secondary_Nav`, and `Tome_Section_Header` */
 	@media (max-width: 800px) {
 		main {
-			/* handle the moved `Library_Secondary_Nav` */
+			/* handle the moved `Docs_Secondary_Nav` */
 			width: 100%;
 			margin-left: 0;
 		}