+++ CHANGELOG.md
@@ -751,7 +751,7 @@
 - rename ts modules that differ only in case from components ([#24](https://github.com/ryanatkn/fuz/pull/24))
 
   - `$lib/contextmenu_helpers.svelte.ts` from `$lib/contextmenu.svelte.ts`
-  - `$lib/library_helpers.svelte.ts` from `$lib/library.svelte.ts`
+  - `$lib/docs_helpers.svelte.ts` from `$lib/library.svelte.ts`
 
 ## 0.100.2